% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_cypher.R
\name{glue_cypher}
\alias{glue_cypher}
\title{Using glue to make a cypher query}
\usage{
glue_cypher(...)
}
\arguments{
\item{...}{One or more character strings}
}
\value{
A string in which the values of variables have been dereferenced
}
\description{
In the same vein as \code{glue_sql}, create a cypher query
using the start \verb{<<} and end \verb{>>} symbols to reference
variables. Will also use the \code{*} symbol to insert a list of
values.
}
