% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_cypher.R
\name{sql_quote_transformer}
\alias{sql_quote_transformer}
\title{SQL quote transformer}
\usage{
sql_quote_transformer(connection, .na)
}
\arguments{
\item{connection}{A dummy DBI connection}

\item{.na}{The string to be evaluated}
}
\description{
Code copied from glue package to mimic \code{glue_sql} but for cypher.
Not meant to be called by itself
}
